Spark 2.4 Sink for Spark Structured Streaming API
