+++ spark-2.4/src/main/scala/org/neo4j/spark/DataSource.scala
@@ -10,14 +10,13 @@ import org.apache.spark.sql.sources.v2.writer.streaming.StreamWriter
 import org.apache.spark.sql.streaming.OutputMode
 import org.apache.spark.sql.types.StructType
 import org.neo4j.spark.reader.Neo4jDataSourceReader
-import org.neo4j.spark.stream.{Neo4jDataSourceStreamReader, Neo4jDataSourceStreamWriter}
+import org.neo4j.spark.stream.Neo4jDataSourceStreamWriter
 import org.neo4j.spark.util.{Neo4jOptions, Validations}
 import org.neo4j.spark.writer.Neo4jDataSourceWriter
 
 import java.util.{Optional, UUID}
 
 class DataSource extends DataSourceV2
-  with MicroBatchReadSupport
   with StreamWriteSupport
   with ReadSupport
   with DataSourceRegister
@@ -46,10 +45,6 @@ class DataSource extends DataSourceV2
            |${Neo4jOptions.SUPPORTED_SAVE_MODES.mkString(",")}
            |""".stripMargin)
     }
-
-  override def createMicroBatchReader(optional: Optional[StructType], s: String, dataSourceOptions: DataSourceOptions): MicroBatchReader = {
-    new Neo4jDataSourceStreamReader(dataSourceOptions, jobId)
-  }
 
   override def createStreamWriter(queryId: String, schema: StructType, mode: OutputMode, options: DataSourceOptions): StreamWriter = {
     new Neo4jDataSourceStreamWriter(jobId, queryId, schema, options)
