+++ spark-2.4/src/main/scala/org/neo4j/spark/stream/Neo4jDataSourceStreamWriter.scala
@@ -25,16 +25,35 @@ class Neo4jDataSourceStreamWriter(private val jobId: String,
 
   var driverCache = new DriverCache(neo4jOptions.connection, jobId)
 
-  override def commit(epochId: Long, messages: Array[WriterCommitMessage]): Unit = logWarning(s"+++ committed ${queryId}")
+  private def callSchemaService[T](function: SchemaService => T): T = {
+    val driverCache = new DriverCache(neo4jOptions.connection, jobId)
+    val schemaService = new SchemaService(neo4jOptions, driverCache)
+    var hasError = false
+    try {
+      function(schemaService)
+    } catch {
+      case e: Throwable => {
+        hasError = true
+        throw e
+      }
+    } finally {
+      schemaService.close()
+      if (hasError) {
+        driverCache.close()
+      }
+    }
+  }
 
-  override def abort(epochId: Long, messages: Array[WriterCommitMessage]): Unit = logInfo(s"+++ aborted ${queryId}")
+  override def commit(epochId: Long, messages: Array[WriterCommitMessage]): Unit = driverCache.close()
+
+  override def abort(epochId: Long, messages: Array[WriterCommitMessage]): Unit = driverCache.close()
 
   override def createWriterFactory(): DataWriterFactory[InternalRow] = {
       logInfo(s"+++ executing query ${queryId}")
-      val schemaService = new SchemaService(neo4jOptions, driverCache)
-      schemaService.createOptimizations()
-      val scriptResult = schemaService.execute(neo4jOptions.script)
-      schemaService.close()
+      val scriptResult = callSchemaService { schemaService => {
+        schemaService.createOptimizations()
+        schemaService.execute(neo4jOptions.script)
+      }}
       new Neo4jDataWriterFactory(jobId, schema, SaveMode.Append, neo4jOptions, scriptResult)
   }
 }
