+++ spark-2.4/src/test/scala/org/neo4j/spark/DataSourceStreamingReaderTSE.scala
@@ -1,37 +1,33 @@
 package org.neo4j.spark
 
+import org.junit.Assert._
 import org.junit.Test
+
+import java.util.UUID
 
 class DataSourceStreamingReaderTSE extends SparkConnectorScalaBaseTSE {
 
   @Test
-  def testReadStream(): Unit = {
-    val stream = ss.readStream.format(classOf[DataSource].getName)
-      .option("url", "bolt://localhost:7687")
-      .option("authentication.type", "basic")
-      .option("authentication.basic.username", "neo4j")
-      .option("authentication.basic.password", "password")
-      .option("labels", "Movie")
+  def testWriteStream(): Unit = {
+    val rateDf = ss.readStream.format("rate")
+      .option("rowsPerSecond", 1)
       .load()
 
-    val query = stream.writeStream.format("console").start()
-
-    query.awaitTermination()
-  }
-
-  @Test
-  def testWriteStream(): Unit = {
-    val rateDf = ss.readStream.format("rate").load()
-
     val query = rateDf.writeStream.format(classOf[DataSource].getName)
-      .option("url", "bolt://localhost:7687")
-      .option("authentication.type", "basic")
-      .option("authentication.basic.username", "neo4j")
-      .option("authentication.basic.password", "password")
-      .option("labels", "Timestampo")
-      .option("checkpointLocation", "/tmp/checkpoint")
+      .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+      .option("labels", "Timestamp")
+      .option("checkpointLocation", "/tmp/checkpoint/" + UUID.randomUUID().toString)
       .start()
 
-    query.awaitTermination()
+    query.awaitTermination(1000)
+    query.processAllAvailable()
+    query.stop()
+
+    val neo4jRateDf = ss.read.format(classOf[DataSource].getName)
+      .option("url", SparkConnectorScalaSuiteIT.server.getBoltUrl)
+      .option("labels", "Timestamp")
+      .load()
+
+    assertTrue(neo4jRateDf.collectAsList().size() > 0)
   }
 }
